HIP: Paralle_Range and Parallel_MDRange clean std::min calls
Fixing a few issues with std::min calls that won't deduce templates
properly when policy::index_type=long.

+++ core/src/HIP/Kokkos_HIP_Parallel_Range.hpp
@@ -349,7 +349,9 @@ class ParallelReduce<FunctorType, Kokkos::RangePolicy<Traits...>, ReducerType,
       // REQUIRED ( 1 , N , 1 )
       const dim3 block(1, block_size, 1);
       // Required grid.x <= block.y
-      const dim3 grid(std::min(block.y, (nwork + block.y - 1) / block.y), 1, 1);
+      const dim3 grid(std::min(block.y, static_cast<uint32_t>(
+                                            (nwork + block.y - 1) / block.y)),
+                      1, 1);
 
       const int shmem =
           UseShflReduction
